restructured the application
