+++ .gitignore
@@ -4,5 +4,6 @@
 venv/
 env/
 __pycache__/
+instance/
