+++ victoria/models/__init__.py
@@ -1,0 +1,8 @@
+
+from .user import User
+from .post import Post
+
+__all__ = [
+    "User",
+    "Post"
+]
