+++ victoria/routes/__init__.py
@@ -1,0 +1,13 @@
+from .about import about
+from .pagenotfound import pagenotfound
+from .login import login
+from .home import home
+from .register import register
+
+__all__ = [
+    "about",
+    "home",
+    "login",
+    "register",
+    "pagenotfound"
+]
